Update nested route for UI profile
