+++ src/constant/path.ts
@@ -1,7 +1,10 @@
 const path = {
   login: '/login',
   register: '/register',
-  profile: '/profile',
+  user: '/user',
+  profile: '/user/profile',
+  changePassword: '/user/change-password',
+  purchase: '/user/purchase',
   home: '/',
   logout: '/logout',
   cart: '/cart',
