--- src/pages/Profile/index.ts
@@ -1,3 +1,0 @@
-import Profile from './Profile'
-
-export default Profile
