+++ src/pages/User/pages/ChangePassword/ChangePassword.tsx
@@ -1,0 +1,3 @@
+export default function ChangePassword() {
+  return <div>ChangePassword</div>
+}
