+++ src/pages/User/pages/Purchase/Purchase.tsx
@@ -1,0 +1,3 @@
+export default function Purchase() {
+  return <div>Purchase</div>
+}
