+++ src/useRouteElement.tsx
@@ -4,12 +4,15 @@ import Login from './pages/Login'
 import Register from './pages/Register'
 import RegisterLayout from './layouts/RegisterLayout'
 import Home from './pages/Home'
-import Profile from './pages/Profile'
 import { AppContext } from './Contexts/app.context'
 import { useContext } from 'react'
 import path from './constant/path'
 import ProductDetail from './pages/ProductDetail/ProductDetail'
 import Cart from './pages/Cart/Cart'
+import UserLayout from './pages/User/layouts/UserLayout/UserLayout'
+import Profile from './pages/User/pages/Profile/Profile'
+import ChangePassword from './pages/User/pages/ChangePassword/ChangePassword'
+import Purchase from './pages/User/pages/Purchase/Purchase'
 
 function ProtectedRoute() {
   const { isAuthenticated } = useContext(AppContext)
@@ -49,12 +52,26 @@ export default function useRouteElement() {
       element: <ProtectedRoute />,
       children: [
         {
-          path: path.profile,
+          path: path.user,
           element: (
             <Home>
-              <Profile />
+              <UserLayout />
             </Home>
-          )
+          ),
+          children: [
+            {
+              path: path.profile,
+              element: <Profile />
+            },
+            {
+              path: path.changePassword,
+              element: <ChangePassword />
+            },
+            {
+              path: path.purchase,
+              element: <Purchase />
+            }
+          ]
         },
         {
           path: path.cart,
